backend/requirements.txt
attrs==23.1.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.6.3
bs4==0.0.1
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
Flask==3.0.3
flask-sitemapper
Flask-SQLAlchemy==3.1.1
google-ai-generativelanguage
google-api-core==2.12.0
google-auth==2.23.4
google-generativeai
googleapis-common-protos==1.61.0
greenlet==3.0.0
grpcio==1.59.2
grpcio-status==1.59.2
h11==0.14.0
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
outcome==1.3.0.post0
proto-plus==1.22.3
protobuf==4.25.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PySocks==1.7.1
requests==2.31.0
rsa==4.9
selenium==4.15.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.22
tqdm==4.66.1
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.8.0
urllib3==2.0.7
Werkzeug==3.0.1
wsproto==1.2.0
python-dotenv
gunicorn
